Wireless-Controller/requirements.txt
kivy
kivymd
pyserial
